src/requirements.txt
albumentations==1.4.6
boto3==1.35.23
ipython==8.12.3
matplotlib==3.8.4
numpy==2.1.1
Pillow==10.4.0
psutil==5.9.0
torch==2.4.0.post101
torchvision==0.17.1
tqdm==4.66.2
zipfile